refactor: consolidate auth, collapse API layer, inject DI, split profile/search, remove dead code
- Merge 3 auth adapters into lib/auth.ts with allowAdmin option
- Collapse 12 API files into 4 domain modules + shared buildQuery
- Add dependency injection pattern to server actions (reviewBookLogic, registerActivityLogic, borrowBookLogic)
- Split 173-line actions/profile.ts into dashboard/settings/favorites
- Extract pure parse/serialize functions from useSearch into lib/search-utils.ts
- Delete dead lib/isAdmin.ts, lib/server/require-user.ts, hooks/use-disclosure.ts
- Remove faker import from ActivityInformations.tsx
- tsc --noEmit passes
