+++ actions/admin/create-first-user.ts
@@ -1,7 +1,7 @@
 'use server'
 import config from '@/payload.config'
 import { getPayload } from 'payload'
-import { setPayloadTokenCookie } from '@/lib/server/set-cookie'
+import { setPayloadTokenCookie } from '@/lib/auth'
 
 interface CreateFirstUserResult {
   ok: boolean
