+++ app/(frontend)/activities/page.tsx
@@ -6,7 +6,7 @@ import ListingContent from '@/components/listing/ListingContent'
 import ListingToolbar from '@/components/listing/listing-toolbar/ListingToolbar'
 import ListingRenderer from '@/components/listing/ListingRenderer'
 import ActivityCard from './_components/ActivityCard'
-import { useGetActivitiesQuery } from '@/lib/apis/activities/queries'
+import { useGetActivitiesQuery } from '@/lib/apis/activities'
 import { ActivitySearchParams, ActivityType } from '@/interfaces/activities.interfaces'
 import { useSearch } from '@/hooks/use-search'
 import { Tag } from 'lucide-react'
