+++ app/(frontend)/articles/page.tsx
@@ -10,7 +10,7 @@ import { useSearch } from '@/hooks/use-search'
 import { ArticleSearchParams, ArticleType } from '@/interfaces/articles.interfaces'
 import EmptyData from '@/components/common/EmptyData'
 import ErrorData from '@/components/common/ErrorData'
-import { useGetArticlesQuery } from '@/lib/apis/articles/queries'
+import { useGetArticlesQuery } from '@/lib/apis/articles'
 import { articleTypesConfigArray } from '@/utils/constants/articles'
 import ArticleCardSkeleton from '@/components/ui/landing/ArticleCardSkeleton'
 import { Tag } from 'lucide-react'
