+++ app/(frontend)/page.tsx
@@ -8,9 +8,9 @@ import ActivityCard from '@/components/ui/activityCard'
 import CTASection from '@/components/ui/CTASection'
 import Image from 'next/image'
 import { motion } from 'motion/react'
-import { useGetBooksQuery } from '@/lib/apis/books/queries'
-import { useGetArticlesQuery } from '@/lib/apis/articles/queries'
-import { useGetActivitiesQuery } from '@/lib/apis/activities/queries'
+import { useGetBooksQuery } from '@/lib/apis/books'
+import { useGetArticlesQuery } from '@/lib/apis/articles'
+import { useGetActivitiesQuery } from '@/lib/apis/activities'
 import { BookCategory } from '@/interfaces/books.interfaces'
 import { Media } from '@/payload-types'
 import { getImageUrl } from '@/utils/image-utils'
