+++ app/(frontend)/profile/_components/ProfileFavoritesGrid.tsx
@@ -7,7 +7,7 @@ import { Card, CardContent } from '@/components/ui/card'
 import { Button } from '@/components/ui/button'
 import { Badge } from '@/components/ui/badge'
 import type { Book, BookFavorite, Media } from '@/payload-types'
-import { removeBookFavorite } from '@/actions/profile'
+import { removeBookFavorite } from '@/actions/profile/favorites'
 import { toast } from 'sonner'
 import { useRouter } from 'next/navigation'
 import EmptyData from '@/components/common/EmptyData'
