+++ hooks/use-search.ts
@@ -2,7 +2,14 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import type { SearchStorageAdapter } from '../interfaces/search.interfaces'
 import { createSearchParamsAdapter } from '@/lib/adapters/search-params.adapter'
-import { isValid, format, parse } from 'date-fns'
+import {
+  type FieldOverride,
+  type FieldOverrides,
+  type PrimitiveType,
+  parseSearchParams,
+  serializeSearchParams,
+  detectType,
+} from '@/lib/search-utils'
 
 function useDebounce<T>(value: T, delay: number): T {
   const [debounced, setDebounced] = useState(value)
@@ -13,17 +20,7 @@ function useDebounce<T>(value: T, delay: number): T {
   return debounced
 }
 
-export type PrimitiveType = 'string' | 'number' | 'boolean' | 'date' | 'enum' | 'array'
-
-export type FieldOverride<T> = {
-  parse?: (value: string) => T
-  serialize?: (value: T) => string
-  type?: PrimitiveType
-}
-
-export type FieldOverrides<T extends object> = {
-  [K in keyof T]?: FieldOverride<T[K]>
-}
+export type { FieldOverride, FieldOverrides, PrimitiveType }
 
 type ResetTransform<T> = (initial: T, current: T) => T
 
@@ -46,44 +43,6 @@ export type SetValue<T> = <K extends keyof T>(
 
 type SetValues<T> = (payload: Partial<T>, options?: { debounced?: boolean }) => void
 
-const defaultParsers: Record<PrimitiveType, (v: string) => unknown> = {
-  string: (v) => v,
-
-  number: (v) => {
-    const n = Number(v)
-    return isNaN(n) ? undefined : n
-  },
-
-  boolean: (v) => v === 'true' || v === '1',
-
-  date: (v) => {
-    const d = parse(v, 'yyyy-MM-dd', new Date())
-    return isValid(d) ? d : undefined
-  },
-
-  enum: (v) => v,
-
-  array: (v) => v.split(','),
-}
-
-const defaultSerializers: Record<PrimitiveType, (v: unknown) => string> = {
-  string: (v) => String(v),
-  number: (v) => String(v),
-  boolean: (v) => ((v as boolean) ? 'true' : 'false'),
-  date: (v) => (v instanceof Date && isValid(v) ? format(v, 'yyyy-MM-dd') : ''),
-  enum: (v) => String(v ?? ''),
-  array: (v) => (Array.isArray(v) ? v.join(',') : ''),
-}
-
-const detectType = (value: unknown): PrimitiveType => {
-  if (value instanceof Date) return 'date'
-  if (Array.isArray(value)) return 'array'
-  if (typeof value === 'boolean') return 'boolean'
-  if (typeof value === 'number') return 'number'
-  if (typeof value === 'string') return 'string'
-  return 'enum'
-}
-
 export const useSearch = <T extends object>({
   initialValues,
   fieldsConfig,
@@ -103,29 +62,7 @@ export const useSearch = <T extends object>({
 
   const parseStored = useCallback((): T => {
     const stored = storage.read()
-
-    const parsed: Partial<T> = {}
-
-    for (const key of keys) {
-      const raw = stored[keyName(key)]
-      const defaultValue = initialValues[key]
-
-      if (!raw) {
-        parsed[key] = defaultValue
-        continue
-      }
-
-      const cfg = fieldsConfig?.[key]
-      const type = cfg?.type ?? detectType(defaultValue)
-
-      const parser = cfg?.parse ?? defaultParsers[type]
-
-      const parsedValue = parser(raw)
-
-      parsed[key] = parsedValue !== undefined ? (parsedValue as T[keyof T]) : defaultValue
-    }
-
-    return { ...initialValues, ...parsed }
+    return parseSearchParams(stored, initialValues, fieldsConfig, keyName)
   }, [initialValues, fieldsConfig, storage, keyName])
 
   const [values, setInternalValues] = useState<T>(() => parseStored())
@@ -198,29 +135,7 @@ export const useSearch = <T extends object>({
   )
 
   useEffect(() => {
-    const serialized: Record<string, string> = {}
-
-    for (const key of keys) {
-      const val = searchValues[key]
-      const defaultValue = initialValues[key]
-
-      if (
-        val === undefined ||
-        val === null ||
-        val === '' ||
-        JSON.stringify(val) === JSON.stringify(defaultValue)
-      ) {
-        continue
-      }
-
-      const cfg = fieldsConfig?.[key]
-      const type = cfg?.type ?? detectType(val)
-
-      const serializer = cfg?.serialize ?? defaultSerializers[type]
-
-      serialized[keyName(key)] = serializer(val)
-    }
-
+    const serialized = serializeSearchParams(searchValues, initialValues, fieldsConfig, keyName)
     storage.write(serialized)
   }, [searchValues, keyName])
