+++ lib/apis/articles.ts
@@ -1,0 +1,60 @@
+import { useQuery } from '@tanstack/react-query'
+import { httpClient } from './http-client'
+import { PaginatedDocs, Where } from 'payload'
+import { Article } from '@/payload-types'
+import { ArticleSearchParams } from '@/interfaces/articles.interfaces'
+import { stringify } from 'qs-esm'
+
+export const articlesKeys = {
+  list: (params?: ArticleSearchParams) => ['articles', 'list', params] as const,
+  detail: (id: string) => ['articles', 'detail', id] as const,
+}
+
+export async function fetchArticles(params?: ArticleSearchParams) {
+  const andFilters: Where[] = []
+
+  if (params?.types) {
+    andFilters.push({ type: { in: params.types } })
+  }
+
+  if (params?.search) {
+    andFilters.push({
+      or: [
+        { title: { contains: params.search } },
+        { description: { contains: params.search } },
+        { author: { contains: params.search } },
+      ],
+    })
+  }
+
+  const query: Where = andFilters.length > 0 ? { and: andFilters } : {}
+
+  const queryString = stringify(
+    {
+      where: query,
+      page: params?.page || 1,
+      limit: params?.limit || 10,
+      sort: '-createdAt',
+    },
+    { addQueryPrefix: true, encodeValuesOnly: true },
+  )
+  return httpClient.get<PaginatedDocs<Article>>(`/articles${queryString}`)
+}
+
+export async function fetchArticleById(id: string) {
+  return httpClient.get<Article>(`/articles/${id}`)
+}
+
+export function useGetArticlesQuery(params?: ArticleSearchParams) {
+  return useQuery({
+    queryKey: articlesKeys.list(params),
+    queryFn: () => fetchArticles(params),
+  })
+}
+
+export function useGetArticleByIdQuery(id: string) {
+  return useQuery({
+    queryKey: articlesKeys.detail(id),
+    queryFn: () => fetchArticleById(id),
+  })
+}
