--- lib/apis/auth/keys.ts
@@ -1,3 +1,0 @@
-export const authKeys = {
-  profile: ['profile'] as const,
-}
