--- lib/isAdmin.ts
@@ -1,2 +1,0 @@
-export const isAdmin = (user: { collection?: string } | null | undefined) =>
-  user?.collection === 'admins'
